feat : debateBoard update
api Docs 명세서에 따른 백엔드를 점검 중에 api Docs에 변한 부분과 추가된 데이터가 있어 이를 반영했습니다.

+++ backend/src/main/java/com/ssafy/backend/db/entity/Reply.java
@@ -26,7 +26,7 @@ public class Reply  {
     boolean hidden;
 
     @Column(name="parent_no")
-    int parentNo;
+    long parentNo;
     
     //연관
     //게시판
